amlvsink: sync roxton/sandia changes
PD#SWPL-123196

Problem:
Backport projects code

Solution:
Backport projects code

Verify:
T5W

+++ src/v4l-dec.h
@@ -87,7 +87,7 @@ struct capture_buffer** v4l_setup_capture_port (int fd, uint32_t *buf_cnt,
 int recycle_output_port_buffer (int fd, struct output_buffer **ob, uint32_t num);
 int recycle_capture_port_buffer (int fd, struct capture_buffer **cb, uint32_t num);
 
-int v4l_dec_dw_config(int fd, uint32_t fmt, uint32_t dw_mode, bool low_latency, bool only_2k, float frame_rate);
+int v4l_dec_dw_config(int fd, uint32_t fmt, uint32_t dw_mode, bool low_latency, bool only_2k, int frame_rate);
 int v4l_dec_config(int fd, bool secure, uint32_t fmt, uint32_t dw_mode,
     struct hdr_meta *hdr, bool is_2k_only, float frame_rate, bool disable_dw_scale);
 int v4l_set_output_format(int fd, uint32_t format, int w, int h, bool only_2k);
